fixes https://luceeserver.atlassian.net/browse/LDEV-878
