+++ core/src/main/java/lucee/runtime/functions/arrays/ArrayNew.java
@@ -28,17 +28,24 @@ import lucee.runtime.ext.function.BIF;
 import lucee.runtime.op.Caster;
 import lucee.runtime.type.Array;
 import lucee.runtime.type.ArrayImpl;
+import lucee.runtime.type.util.ArrayUtil;
 
 public final class ArrayNew extends BIF {
 
 	private static final long serialVersionUID = -5923269433550568279L;
 
 	public static Array call(PageContext pc) throws ExpressionException  {
-		return new ArrayImpl(1);
+		return new ArrayImpl();
 	}
 	
-	public static Array call(PageContext pc , double number) throws ExpressionException {
-		return new ArrayImpl((int)number);
+	public static Array call(PageContext pc , double dimension) throws ExpressionException {
+		return ArrayUtil.getInstance((int)dimension);
+	}
+	
+	public static Array call(PageContext pc , double dimension, boolean isSynchronized) throws ExpressionException {
+		if(dimension>1)
+			return ArrayUtil.getInstance((int)dimension);
+		return new ArrayImpl(isSynchronized);
 	}
 	
 	@Override
