+++ core/src/main/java/lucee/runtime/functions/arrays/ArrayReverse.java
@@ -28,6 +28,7 @@ import lucee.runtime.ext.function.BIF;
 import lucee.runtime.op.Caster;
 import lucee.runtime.type.Array;
 import lucee.runtime.type.ArrayImpl;
+import lucee.runtime.type.util.ArrayUtil;
 
 
 public final class ArrayReverse extends BIF {
@@ -35,7 +36,7 @@ public final class ArrayReverse extends BIF {
 	private static final long serialVersionUID = 5418304787535992180L;
 
 	public static Array call(PageContext pc , Array array) throws ExpressionException {
-		Array rev=new ArrayImpl(array.getDimension());
+		Array rev=ArrayUtil.getInstance(array.getDimension());
 		int len=array.size();
 		for(int i=0;i<len;i++) {
 			try {
