+++ core/src/main/java/lucee/runtime/functions/arrays/ArraySlice.java
@@ -28,6 +28,7 @@ import lucee.runtime.ext.function.BIF;
 import lucee.runtime.op.Caster;
 import lucee.runtime.type.Array;
 import lucee.runtime.type.ArrayImpl;
+import lucee.runtime.type.util.ArrayUtil;
 
 public final class ArraySlice extends BIF {
 	
@@ -60,7 +61,7 @@ public final class ArraySlice extends BIF {
 	}
 	
 	private static Array get(Array arr, int from, int to) throws PageException {
-		Array rtn=new ArrayImpl(arr.getDimension());
+		Array rtn=ArrayUtil.getInstance(arr.getDimension());
 		int[] keys=arr.intKeys();
 		for(int i=0;i<keys.length;i++) {
 			int key=keys[i];
