+++ core/src/main/java/lucee/runtime/type/KeyImpl.java
@@ -294,14 +294,14 @@ public class KeyImpl implements Collection.Key,Castable,Comparable,Externalizabl
 	public static Array toUpperCaseArray(Key[] keys) {
 		ArrayImpl arr=new ArrayImpl();
 		for(int i=0;i<keys.length;i++) {
-			arr._append(((KeyImpl)keys[i]).getUpperString());
+			arr.appendEL(((KeyImpl)keys[i]).getUpperString());
 		}
 		return arr;
 	}
 	public static Array toLowerCaseArray(Key[] keys) {
 		ArrayImpl arr=new ArrayImpl();
 		for(int i=0;i<keys.length;i++) {
-			arr._append(((KeyImpl)keys[i]).getLowerString());
+			arr.appendEL(((KeyImpl)keys[i]).getLowerString());
 		}
 		return arr;
 	}
@@ -309,7 +309,7 @@ public class KeyImpl implements Collection.Key,Castable,Comparable,Externalizabl
 	public static Array toArray(Key[] keys) {
 		ArrayImpl arr=new ArrayImpl();
 		for(int i=0;i<keys.length;i++) {
-			arr._append(((KeyImpl)keys[i]).getString());
+			arr.appendEL(((KeyImpl)keys[i]).getString());
 		}
 		return arr;
 	}
