+++ core/src/main/java/lucee/runtime/type/util/ArrayUtil.java
@@ -27,6 +27,7 @@ import java.util.List;
 import java.util.ListIterator;
 import java.util.Map;
 import java.util.Set;
+import java.util.Map.Entry;
 
 import lucee.commons.lang.ArrayUtilException;
 import lucee.commons.lang.ComparatorUtil;
@@ -40,9 +41,15 @@ import lucee.runtime.exp.ExpressionException;
 import lucee.runtime.exp.PageException;
 import lucee.runtime.op.Caster;
 import lucee.runtime.op.Decision;
+import lucee.runtime.op.Duplicator;
 import lucee.runtime.op.Operator;
+import lucee.runtime.op.ThreadLocalDuplication;
 import lucee.runtime.type.Array;
+import lucee.runtime.type.ArrayClassic;
+import lucee.runtime.type.ArrayImpl;
+import lucee.runtime.type.Collection;
 import lucee.runtime.type.QueryColumn;
+import lucee.runtime.type.Collection.Key;
 import lucee.runtime.type.comparator.SortRegister;
 
 /**
@@ -52,6 +59,12 @@ public final class ArrayUtil {
     
     public static final Object[] OBJECT_EMPTY = new Object[]{};
 	
+    public static Array getInstance(int dimension) throws ExpressionException {
+    	if(dimension>1) return new ArrayClassic(dimension);
+        return new ArrayImpl();
+    }
+    
+    
     /**
      * trims all value of a String Array
      * @param arr
@@ -991,5 +1004,4 @@ public final class ArrayUtil {
 			list.add(arr[i]);
 		}
 	}
-
 }
